Orville.PostgreSQL.Execution

Description

You can import Orville.PostgreSQL.Execution to get access to all the execution-related functions and types exported by the modules below. This includes a number of lowel-level items not exported by Orville.PostgreSQL that give you more control (and therefore responsibility) over how the SQL is executed.

Since: 1.0.0.0

High level modules for most common tasks

module Orville.PostgreSQL.Execution.EntityOperations

module Orville.PostgreSQL.Execution.SelectOptions

module Orville.PostgreSQL.Execution.Sequence

module Orville.PostgreSQL.Execution.Transaction

Mid-level modules with more flexibility for handling less common queries

module Orville.PostgreSQL.Execution.Select

module Orville.PostgreSQL.Execution.Insert

module Orville.PostgreSQL.Execution.Update

module Orville.PostgreSQL.Execution.Delete

module Orville.PostgreSQL.Execution.Cursor

module Orville.PostgreSQL.Execution.ReturningOption

Low-level modules for executing and decoding SQL expressions yourself

module Orville.PostgreSQL.Execution.Execute

module Orville.PostgreSQL.Execution.ExecutionResult

module Orville.PostgreSQL.Execution.QueryType