Orville.PostgreSQL.Execution.EntityOperations

Description

Entry-level functions for executing based CRUD operations on entity tables. This are the functions that you will use most often for interacting with tables.

Since: 1.0.0.0

Documentation

insertEntity :: MonadOrville m => TableDefinition key writeEntity readEntity -> writeEntity -> m ()

Inserts a entity into the specified table.

Since: 1.0.0.0

insertEntityAndReturnRowCount :: MonadOrville m => TableDefinition key writeEntity readEntity -> writeEntity -> m Int

Inserts a entity into the specified table. Returns the number of rows affected by the query.

Since: 1.0.0.0

insertAndReturnEntity :: MonadOrville m => TableDefinition key writeEntity readEntity -> writeEntity -> m readEntity

Inserts a entity into the specified table, returning the data inserted into the database.

You can use this function to obtain any column values filled in by the database, such as auto-incrementing ids.

Since: 1.0.0.0

insertEntities :: MonadOrville m => TableDefinition key writeEntity readEntity -> NonEmpty writeEntity -> m ()

Inserts a non-empty list of entities into the specified table.

Since: 1.0.0.0

insertAndReturnEntities :: MonadOrville m => TableDefinition key writeEntity readEntity -> NonEmpty writeEntity -> m [readEntity]

Inserts a non-empty list of entities into the specified table, returning the data that was inserted into the database.

You can use this function to obtain any column values filled in by the database, such as auto-incrementing ids.

Since: 1.0.0.0

insertEntitiesAndReturnRowCount :: MonadOrville m => TableDefinition key writeEntity readEntity -> NonEmpty writeEntity -> m Int

Inserts a non-empty list of entities into the specified table. Returns the number of rows affected by the query.

Since: 1.0.0.0

updateEntity :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> writeEntity -> m ()

Updates the row with the given key in with the data given by writeEntity.

Since: 1.0.0.0

updateEntityAndReturnRowCount :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> writeEntity -> m Int

Updates the row with the given key in with the data given by writeEntity. Returns the number of rows affected by the query.

Since: 1.0.0.0

updateAndReturnEntity :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> writeEntity -> m (Maybe readEntity)

Updates the row with the given key in with the data given by writeEntity, returning updated row from the database. If no row matches the given key, Nothing will be returned.

You can use this function to obtain any column values computer by the database during update, including columns with triggers attached to them.

Since: 1.0.0.0

updateFields :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> NonEmpty SetClause -> Maybe BooleanExpr -> m ()

Applies the given SetClauses to the rows in the table that match the given where condition. The easiest way to construct a SetClause is via the setField function (also exported as .:=).

Since: 1.0.0.0

updateFieldsAndReturnEntities :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> NonEmpty SetClause -> Maybe BooleanExpr -> m [readEntity]

Like updateFields, but uses a RETURNING clause to return the updated version of any rows that were affected by the update.

Since: 1.0.0.0

updateFieldsAndReturnRowCount :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> NonEmpty SetClause -> Maybe BooleanExpr -> m Int

Applies the given SetClauses to the rows in the table that match the given where condition. The easiest way to construct a SetClause is via the setField function (also exported as .:=). Returns the number of rows affected by the query.

Since: 1.0.0.0

deleteEntity :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> m ()

Deletes the row with the given key.

Since: 1.0.0.0

deleteEntityAndReturnRowCount :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> m Int

Deletes the row with the given key. Returns the number of rows affected by the query.

Since: 1.0.0.0

deleteAndReturnEntity :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> m (Maybe readEntity)

Deletes the row with the given key, returning the row that was deleted. If no row matches the given key, Nothing is returned.

Since: 1.0.0.0

deleteEntities :: MonadOrville m => TableDefinition key writeEntity readEntity -> Maybe BooleanExpr -> m ()

Deletes all rows in the given table that match the where condition.

Since: 1.0.0.0

deleteEntitiesAndReturnRowCount :: MonadOrville m => TableDefinition key writeEntity readEntity -> Maybe BooleanExpr -> m Int

Deletes all rows in the given table that match the where condition. Returns the number of rows affected by the query.

Since: 1.0.0.0

deleteAndReturnEntities :: MonadOrville m => TableDefinition key writeEntity readEntity -> Maybe BooleanExpr -> m [readEntity]

Deletes all rows in the given table that match the where condition, returning the rows that were deleted.

Since: 1.0.0.0

findEntitiesBy :: MonadOrville m => TableDefinition key writeEntity readEntity -> SelectOptions -> m [readEntity]

Finds all the entities in the given table according to the specified SelectOptions, which may include where conditions to match, ordering specifications, etc.

Since: 1.0.0.0

findFirstEntityBy :: MonadOrville m => TableDefinition key writeEntity readEntity -> SelectOptions -> m (Maybe readEntity)

Like 'findEntitiesBy, but adds a 'LIMIT 1' to the query and then returns the first item from the list. Usually when you use this you will want to provide an order by clause in the SelectOptions because the database will not guarantee ordering.

Since: 1.0.0.0

findEntity :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> key -> m (Maybe readEntity)

Finds a single entity by the table's primary key value.

Since: 1.0.0.0

findEntities :: MonadOrville m => TableDefinition (HasKey key) writeEntity readEntity -> NonEmpty key -> m [readEntity]

Finds multiple entities by the table's primary key.

Since: 1.0.0.0