Orville.PostgreSQL.Execution.ExecutionResult

Description

Since: 1.0.0.0

Documentation

class ExecutionResult result where

ExecutionResult is a common interface for types that represent a result set returned from the database. For real, live database interactions this the concrete type will be a Result, but the FakeLibPQResult may be useful as well if you are writing custom code for decoding result sets and want to test aspects of the decoding that don't require a real database.

Since: 1.0.0.0

Methods

maxRowNumber :: result -> IO (Maybe Row)

maxColumnNumber :: result -> IO (Maybe Column)

columnName :: result -> Column -> IO (Maybe ByteString)

getValue :: result -> Row -> Column -> IO SqlValue

newtype Column

A trivial wrapper for Int to help keep track of column vs row number

Since: 1.0.0.0

Constructors

Column Int

newtype Row

A trivial wrapper for Int to help keep track of column vs row number

Since: 1.0.0.0

Constructors

Row Int

readRows :: ExecutionResult result => result -> IO [[(Maybe ByteString, SqlValue)]]

Read the rows of an ExecutionResult a list of column name, SqlValue pairs. You're almost always better off using a SqlMarshaller instead, but this function is provided for cases where you really want to decode the rows yourself but don't want to use the ExecutionResult api to read each row of each column directly.

Since: 1.0.0.0

data FakeLibPQResult

FakeLibPQResult provides a fake, in memory implementation of ExecutionResult. This is mostly useful for writing automated tests that can assume a result set has been loaded and just need to test decoding the results.

Since: 1.0.0.0

mkFakeLibPQResult

Arguments

:: [ByteString]
The column names for the result set

-> [[SqlValue]]
The row data for the result set

-> FakeLibPQResult

Constructs a FakeLibPQResult. The column names given as associated with the values for each row by their position in list. Any missing values (e.g. because a row is shorter than the heeader list) are treated as a SQL Null value.

Since: 1.0.0.0