Orville.PostgreSQL.Execution.QueryType

Description

Since: 1.0.0.0

Documentation

data QueryType

A simple categorization of SQL queries that is used to provide a hint to user callbacks about what kind of query is being run.

See addSqlExecutionCallback

Since: 1.0.0.0

Constructors

SelectQuery

InsertQuery

UpdateQuery

DeleteQuery

DDLQuery

CursorQuery

OtherQuery