Orville.PostgreSQL.Execution.ReturningOption

Description

Since: 1.0.0.0

Documentation

data ReturningOption clause where

Specifies whether or not a RETURNING clause should be included when a query expression is built. This type is found as a parameter on a number of the query building functions related to TableDefinition. @since 1.0.0.0

Constructors

WithReturning :: ReturningOption ReturningClause

WithoutReturning :: ReturningOption NoReturningClause

data ReturningClause

A tag, used with ReturningOption to indicate a SQL Returning clause.

Since: 1.0.0.0

data NoReturningClause

A tag, used with ReturningOption to indicate no SQL Returning clause.

Since: 1.0.0.0