Orville.PostgreSQL.Execution.Sequence

Description

Interactions to work with database sequence values on the Haskell side. Including inspection of the current and next values in the sequence as well as updating a sequence to a given value.

Since: 1.0.0.0

Documentation

sequenceNextValue :: MonadOrville m => SequenceDefinition -> m Int64

Fetches the next value from a sequence via the PostgreSQL nextval function.

Since: 1.0.0.0

sequenceCurrentValue :: MonadOrville m => SequenceDefinition -> m Int64

Fetches the current value from a sequence via the PostgreSQL currval function.

Since: 1.0.0.0

sequenceSetValue :: MonadOrville m => SequenceDefinition -> Int64 -> m Int64

Sets the current value from a sequence via the PostgreSQL setval function.

Since: 1.0.0.0