Orville.PostgreSQL.Expr.BinaryOperator

Description

Provides a type representing sql operators with exactly two arguments, as well as values of that type for many common operators.

Since: 1.0.0.0

Documentation

data BinaryOperator

Type to represent any SQL operator of two arguments. E.G.

AND

BinaryOperator provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

binaryOperator :: String -> BinaryOperator

Construct a binary operator, note that this does not include any check to determine if the operator is valid either by being a native SQL operator, or a custom defined operator in the database.

Since: 1.0.0.0

binaryOpExpression :: SqlExpression sql => BinaryOperator -> ValueExpression -> ValueExpression -> sql

Apply a binary operator to two ValueExpressions resulting in some SqlExpression. Note that this does *NOT* extend typechecking to the ValueExpressions being used with the BinaryOperator. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

equalsOp :: BinaryOperator

The SQL equal binary operator.

Since: 1.0.0.0

notEqualsOp :: BinaryOperator

The SQL not equal binary operator.

Since: 1.0.0.0

greaterThanOp :: BinaryOperator

The SQL strictly greater than binary operator.

Since: 1.0.0.0

lessThanOp :: BinaryOperator

The SQL strictly less than binary operator.

Since: 1.0.0.0

greaterThanOrEqualsOp :: BinaryOperator

The SQL greater than or equal binary operator.

Since: 1.0.0.0

lessThanOrEqualsOp :: BinaryOperator

The SQL less than or equal binary operator.

Since: 1.0.0.0

likeOp :: BinaryOperator

The SQL LIKE binary operator.

Since: 1.0.0.0

iLikeOp :: BinaryOperator

The SQL ILIKE binary operator.

Since: 1.0.0.0

orOp :: BinaryOperator

The SQL logical or binary operator.

Since: 1.0.0.0

andOp :: BinaryOperator

The SQL logical and binary operator.

Since: 1.0.0.0

plusOp :: BinaryOperator

The SQL + binary operator.

Since: 1.0.0.0

minusOp :: BinaryOperator

The SQL - binary operator.

Since: 1.0.0.0

multiplicationOp :: BinaryOperator

The SQL * binary operator.

Since: 1.0.0.0

divisionOp :: BinaryOperator

The SQL / binary operator.

Since: 1.0.0.0

moduloOp :: BinaryOperator

The SQL % binary operator.

Since: 1.0.0.0

exponentiationOp :: BinaryOperator

The SQL ^ binary operator.

Since: 1.0.0.0

bitwiseAndOp :: BinaryOperator

The SQL bitwise and (a.k.a &) binary operator.

Since: 1.0.0.0

bitwiseOrOp :: BinaryOperator

The SQL bitwise or (a.k.a |) binary operator.

Since: 1.0.0.0

bitwiseXorOp :: BinaryOperator

The SQL bitwise exclusive or (a.k.a #) binary operator.

Since: 1.0.0.0

bitwiseShiftLeftOp :: BinaryOperator

The SQL bitwise left shift (a.k.a <<) binary operator.

Since: 1.0.0.0

bitwiseShiftRightOp :: BinaryOperator

The SQL bitwise right shift (a.k.a >>) binary operator.

Since: 1.0.0.0