Orville.PostgreSQL.Expr.ColumnDefinition

Description

Since: 1.0.0.0

Documentation

data ColumnDefinition

Represent a complete definition of a column. E.G.

foo INTEGER

ColumnDefinition provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

columnDefinition

Arguments

:: ColumnName
The name the resulting column should have.

-> DataType
The SQL type of the column

-> Maybe ColumnConstraint
The constraint on the column if any

-> Maybe ColumnDefault
The default value for the column if any

-> ColumnDefinition

Smart constructor for ensuring a ColumnDefinition is setup correctly.

Since: 1.0.0.0

data ColumnConstraint

Represent constraints, such as nullability, on a column. E.G.

NOT NULL

ColumnConstraint provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

notNullConstraint :: ColumnConstraint

Express that a column may not contain NULL.

Since: 1.0.0.0

nullConstraint :: ColumnConstraint

Express that a column may contain NULL.

Since: 1.0.0.0

data ColumnDefault

Represents the default value of a column. E.G.

now()

ColumnDefault provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

columnDefault :: ValueExpression -> ColumnDefault

Given a ValueExpression use that as a ColumnDefault. This is the preferred path to creating a column default. Note that it is up to the caller to ensure the ValueExpression makes sense for the resulting ColumnDefinition this will be a part of.

Since: 1.0.0.0