Orville.PostgreSQL.Expr.Count

Description

Since: 1.0.0.0

Documentation

count :: ValueExpression -> ValueExpression

Given a ValueExpression, use it as the argument to the SQL count

Since: 1.0.0.0

countFunction :: FunctionName

The SQL count function

Since: 1.0.0.0

count1 :: ValueExpression

The SQL count(1)

Since: 1.0.0.0

countColumn :: ColumnName -> ValueExpression

Use a given column as the argument to the SQL count

Since: 1.0.0.0