Orville.PostgreSQL.Expr.Cursor

Description

Since: 1.0.0.0

Documentation

data DeclareExpr

DeclareExpr corresponds to the SQL DECLARE statement, for declaring and opening cursors. E.G.

DECLARE FOO CURSOR FOR SELECT * FROM BAR

See PostgreSQL cursor declare documentation for more information.

DeclareExpr provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

declare :: CursorName -> Maybe ScrollExpr -> Maybe HoldExpr -> QueryExpr -> DeclareExpr

A smart constructor for setting up a DeclareExpr. This, along with other functions provided allow to more safely declare a cursor.

Since: 1.0.0.0

data ScrollExpr

ScrollExpr is used to determine if a cursor should be able to fetch nonsequentially. E.G.

NO SCROLL

Note that the default in at least PostgreSQL versions 11-15 is to allow nonsequential fetches under some, but not all, circumstances.

See PostgreSQL cursor declare documentation for more information.

ScrollExpr provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

scroll :: ScrollExpr

Allow a cursor to be used to fetch rows nonsequentially.

Since: 1.0.0.0

noScroll :: ScrollExpr

Only allow a cursor to be used to fetch rows sequentially.

Since: 1.0.0.0

data HoldExpr

HoldExpr is used to determine if a cursor should be available for use after the transaction that created it has been comitted. E.G.

WITH HOLD

See PostgreSQL cursor documentation for more information.

HoldExpr provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

withHold :: HoldExpr

Allow a cursor to be used after the transaction creating it is committed.

Since: 1.0.0.0

withoutHold :: HoldExpr

Do not allow a cursor to be used after the transaction creating it is committed.

Since: 1.0.0.0

data CloseExpr

CloseExpr corresponds to the SQL CLOSE statement. E.G.

CLOSE ALL

See PostgreSQL close documentation for more information.

HoldExpr provides' a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

close :: Either AllCursors CursorName -> CloseExpr

A smart constructor for setting up a CloseExpr either closing all cursors or the given named cursor.

Since: 1.0.0.0

data AllCursors

AllCursors corresponds to the ALL keyword in a CLOSE statement. E.G.

ALL

AllCursors provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

allCursors :: AllCursors

Specify closing all open cursors, for use with a CloseExpr

Since: 1.0.0.0

data FetchExpr

FetchExpr corresponds to the SQL FETCH statement, for retrieving rows from a previously created cursor. E.G.

FETCH NEXT FOO

See PostgreSQL fetch documentation for more information.

FetchExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

fetch :: Maybe CursorDirection -> CursorName -> FetchExpr

Construct a FetchExpr, for a given cursor and optionally a direction to fetch.

Since: 1.0.0.0

data MoveExpr

MoveExpr corresponds to the SQL MOVE statement, for positioning a previously created cursor, without retrieving any rows. E.G.

MOVE NEXT FOO

MoveExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

move :: Maybe CursorDirection -> CursorName -> MoveExpr

Construct a MoveExpr, for a given cursor and optionally a direction to move.

Since: 1.0.0.0

data CursorDirection

CursorDirection corresponds to the direction argument to the SQL FETCH and MOVE statements. E.G.

BACKWARD

See PostgreSQL fetch documentation for more information.

CursorDirection provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

next :: CursorDirection

Specify a direction of the next single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

prior :: CursorDirection

Specify a direction of the prior single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

first :: CursorDirection

Specify a direction of the first single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

last :: CursorDirection

Specify a direction of the last single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

absolute :: Int -> CursorDirection

Specify a direction of the single row at an absolute position within the cursor. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

relative :: Int -> CursorDirection

Specify a direction of the single row relative to the cursor's current position. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

rowCount :: Int -> CursorDirection

Specify a direction of the next n rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

fetchAll :: CursorDirection

Specify a direction of all the next rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

forward :: CursorDirection

Specify a direction of the next single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

forwardCount :: Int -> CursorDirection

Specify a direction of the next n rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

forwardAll :: CursorDirection

Specify a direction of all the next rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

backward :: CursorDirection

Specify a direction of the prior single row. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

backwardCount :: Int -> CursorDirection

Specify a direction of the prior n rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0

backwardAll :: CursorDirection

Specify a direction of all the prior rows. Primarily for use with fetch or move.

See PostgreSQL fetch documentation for more information.

Since: 1.0.0.0