Orville.PostgreSQL.Expr.Delete

Description

Provides a type representing SQL DELETE and construction of that type.

Since: 1.0.0.0

Documentation

data DeleteExpr

Type to represent a SQL delete statement. E.G.

DELETE FROM foo WHERE id < 10

DeleteExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

deleteExpr :: Qualified TableName -> Maybe WhereClause -> Maybe ReturningExpr -> DeleteExpr

Construct a SQL DELETE from a table, optionally limiting with a WhereClause and optionally returning a ReturningExpr.

Since: 1.0.0.0