Orville.PostgreSQL.Expr.GroupBy

Description

Since: 1.0.0.0

Documentation

data GroupByClause

Type to represent a SQL group by clause. E.G.

GROUP BY team_name

GroupByClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

groupByClause :: GroupByExpr -> GroupByClause

Create a full sql GROUP BY clause with the given expression.

Since: 1.0.0.0

data GroupByExpr

Type to represent a SQL group by expression (the part that follows the GROUP BY in sql). E.G.

team_name

GroupByExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

appendGroupByExpr :: GroupByExpr -> GroupByExpr -> GroupByExpr

Combines two GroupByExprs with a comma between them.

Since: 1.0.0.0

groupByExpr :: RawSql -> GroupByExpr

Create a GroupByExpr from some RawSql. Note that it is up to the caller to ensure that the given value can actually be used for a GroupByExpr

Since: 1.0.0.0

groupByColumnsExpr :: NonEmpty ColumnName -> GroupByExpr

Create a GroupByExpr from the given ColumnNames.

Since: 1.0.0.0