Orville.PostgreSQL.Expr.Index

Description

Since: 1.0.0.0

Documentation

data CreateIndexExpr

Type to represent a SQL "CREATE INDEX" statement. E.G.

CREATE INDEX ON table (foo, bar, baz)

CreateIndexExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

createIndexExpr :: IndexUniqueness -> Maybe ConcurrentlyExpr -> Qualified TableName -> NonEmpty ColumnName -> CreateIndexExpr

Construct a SQL CREATE INDEX from an indicator of if the index should be unique, a table, and corresponding collection of ColumnNames.

Since: 1.0.0.0

data IndexUniqueness

Type to represent if an index should be unique.

Since: 1.0.0.0

Constructors

UniqueIndex

NonUniqueIndex

data IndexBodyExpr

Type to represent if the body of an index definition E.G.

(foo, bar)

in

CREATE some_index ON some_table (foo, bar)

IndexBodyExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

indexBodyColumns :: NonEmpty ColumnName -> IndexBodyExpr

Creates an IndexBodyExpr for the given column names. The resulting SQL looks like (column1, column2, ...).

Since: 1.0.0.0

data ConcurrentlyExpr

Type to represent the CONCURRENTLY keyword for index creation. E.G.

CONCURRENTLY

ConcurrentlyExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

concurrently :: ConcurrentlyExpr

The CONCURRENTLY keyword indicates to PostgreSQL that an index should be create concurrently.

Since: 1.0.0.0

data DropIndexExpr

Type to represent a SQL "DROP INDEX" statement. E.G.

DROP INDEX foo

DropIndexExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

dropIndexExpr :: IndexName -> DropIndexExpr

Construct a SQL DROP INDEX for a given IndexName.

Since: 1.0.0.0

createNamedIndexExpr :: IndexUniqueness -> Maybe ConcurrentlyExpr -> Qualified TableName -> IndexName -> IndexBodyExpr -> CreateIndexExpr

Construct a SQL CREATE INDEX from an indicator of if the index should be unique, a table, a name for the index, and some sql representing the rest of the index creation.

Since: 1.0.0.0