Orville.PostgreSQL.Expr.Math

Description

Since: 1.0.0.0

Documentation

plus :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL + to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

minus :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL - to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

multiply :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL * to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

divide :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL / to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

modulo :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL % to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

exponentiate :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL ^ to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

bitwiseAnd :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL & to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

bitwiseOr :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL | to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

bitwiseXor :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL # to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

bitwiseShiftLeft :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL << to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0

bitwiseShiftRight :: ValueExpression -> ValueExpression -> ValueExpression

Apply a SQL >> to the ValueExpressions. It is left to the caller to ensure that the operator makes sense with the arguments being passed.

Since: 1.0.0.0