Orville.PostgreSQL.Expr.Name

Description

Since: 1.0.0.0

Documentation

data TableName

Type to represent a SQL table name. TableName values constructed via the tableName function will be properly escaped as part of the generated SQL. E.G.

"some_table_name"

TableName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

tableName :: String -> TableName

Construct a TableName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data SequenceName

Type to represent a SQL sequence name. SequenceName values constructed via the sequenceName function will be properly escaped as part of the generated SQL. E.G.

"some_sequence_name"

SequenceName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

sequenceName :: String -> SequenceName

Construct a SequenceName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data SchemaName

Type to represent a SQL schema name. SchemaName values constructed via the schemaName function will be properly escaped as part of the generated SQL. E.G.

"some_schema_name"

SchemaName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

schemaName :: String -> SchemaName

Construct a ColumnName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data SavepointName

Type to represent a SQL savepoint name. SavepointName values constructed via the savepointName function will be properly escaped as part of the generated SQL. E.G.

"some_savepoint_name"

SavepointName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

savepointName :: String -> SavepointName

Construct a SavepointName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data Qualified name

Type to represent a qualified SQL name. E.G.

"some_schema_name"."some_table_name"

Qualified provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

qualifyTable :: Maybe SchemaName -> TableName -> Qualified TableName

Optionally qualifies a TableName with a SchemaName. Generally you would want the higher level function tableIdQualifiedName.

Since: 1.0.0.0

qualifySequence :: Maybe SchemaName -> SequenceName -> Qualified SequenceName

Optionally qualifies a SequenceName with a SchemaName. Generally you would want the higher level function sequenceIdQualifiedName.

Since: 1.0.0.0

qualifyColumn :: Maybe SchemaName -> TableName -> ColumnName -> Qualified ColumnName

Qualifies a ColumnName with a TableName and, optionally, a SchemaName. This should be used to refer to the column in SQL queries where a qualified reference is appropriate.

Since: 1.0.0.0

data IndexName

Type to represent a SQL index name. IndexName values constructed via the indexName function will be properly escaped as part of the generated SQL. E.G.

"some_index_name"

IndexName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

indexName :: String -> IndexName

Construct a IndexName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data Identifier

Type to represent a SQL identifier. Identifier values constructed via the identifier function will be properly escaped as part of the generated SQL. E.G.

"some_identifier"

Identifer provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

identifier :: String -> Identifier

Construct an Identifier from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

identifierFromBytes :: ByteString -> Identifier

Construct an Identifier from a ByteString with proper escaping as part of the generated SQL.

Since: 1.0.0.0

class IdentifierExpression name where

This class aids in giving additional polymorphism and so that many different identifiers can be created without being forced to only use the Identifier type.

Since: 1.0.0.0

Methods

toIdentifier :: name -> Identifier

Since: 1.0.0.0

fromIdentifier :: Identifier -> name

Since: 1.0.0.0

data FunctionName

Type to represent a SQL function name. FunctionName values constructed via the functionName function will be properly escaped as part of the generated SQL. E.G.

"some_function_name"

FunctionName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

functionName :: String -> FunctionName

Construct a FunctionName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data CursorName

Type to represent a SQL cursor name. CursorName values constructed via the cursorName function will be properly escaped as part of the generated SQL. E.G.

"some_cursor_name"

CursorName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

cursorName :: String -> CursorName

Construct a CursorName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data ConstraintName

Type to represent a SQL constraint name. ConstraintName values constructed via the constraintName function will be properly escaped as part of the generated SQL. E.G.

"some_constraint_name"

ConstraintName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

constraintName :: String -> ConstraintName

Construct a ConstraintName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0

data ColumnName

Type to represent a SQL column name. ColumnName values constructed via the columnName function will be properly escaped as part of the generated SQL. E.G.

"some_column_name"

ColumnName provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

columnName :: String -> ColumnName

Construct a ColumnName from a String with proper escaping as part of the generated SQL.

Since: 1.0.0.0