Orville.PostgreSQL.Expr.OrderBy

Description

Since: 1.0.0.0

Documentation

data OrderByClause

Type to represent a SQL order by clause. E.G.

ORDER BY foo, bar

OrderByClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

orderByClause :: OrderByExpr -> OrderByClause

data OrderByExpr

Type to represent a SQL order by expression (the part that follows the ORDER BY in sql). E.G.

foo, bar

OrderByExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

appendOrderByExpr :: OrderByExpr -> OrderByExpr -> OrderByExpr

Combines two OrderByExprs with a comma between them.

Since: 1.0.0.0

orderByColumnName :: ColumnName -> OrderByDirection -> OrderByExpr

orderByExpr :: RawSql -> OrderByDirection -> OrderByExpr

Create a OrderByExpr from some RawSql and a OrderByDirection. Note that it is up to the caller to ensure that the given value can actually be used for a OrderByExpr

Since: 1.0.0.0

orderByColumnsExpr :: NonEmpty (ColumnName, OrderByDirection) -> OrderByExpr

Create a OrderByExpr for ColumnName and OrderByDirection pairs, ensuring commas as needed.

Since: 1.0.0.0

data OrderByDirection

Type to represent a SQL order by direction expression. E.G.

ASC

OrderByDirection provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

data NullsOrder

Type to represent the ordering of Null, intended to be used with OrderByDirection.

Since: 1.0.0.0

Constructors

NullsFirst

NullsLast

ascendingOrder :: OrderByDirection

The SQL ASC order direction.

Since: 1.0.0.0

descendingOrder :: OrderByDirection

The SQL DESC order direction.

Since: 1.0.0.0

ascendingOrderWith :: NullsOrder -> OrderByDirection

The SQL ASC order direction with NULLs ordered as given.

Since: 1.0.0.0

descendingOrderWith :: NullsOrder -> OrderByDirection

The SQL DESC order direction with NULLs ordered as given.

Since: 1.0.0.0