Orville.PostgreSQL.Expr.ReturningExpr

Description

Since: 1.0.0.0

Documentation

data ReturningExpr

Type to represent a RETURNING clause in a SQL SELECT statement. E.G.

RETURNING (id)

ReturningExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

returningExpr :: SelectList -> ReturningExpr

Constructs a ReturningExpr that returns the items given in the SelectList. Essentialy this retults RETURNING items

Since: 1.0.0.0