Orville.PostgreSQL.Expr.Select

Description

Since: 1.0.0.0

Documentation

data SelectClause

Type to represent the SELECT part of SQL query E.G.

SELECT

or

SELECT DISTINCT

SelectClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

selectClause :: SelectExpr -> SelectClause

Constructs a SelectClause using the given SelectExpr, which may indicate that this is a DISTINCT select.

Since: 1.0.0.0

data SelectExpr

Type to represent the any expression modifying the SELECT part of a SQL. E.G.

DISTINCT

SelectExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

selectExpr :: Maybe Distinct -> SelectExpr

Constructs a SelectExpr that may or may not make the SELECT distinct, dependending on whether 'Just Distinct' is passed or not.

Since: 1.0.0.0

data Distinct

A simple value type used to indicate that a SELECT should be distinct when constructing a SelectExpr

Since: 1.0.0.0

Constructors

Distinct