Orville.PostgreSQL.Expr.TableConstraint

Description

Since: 1.0.0.0

Documentation

data TableConstraint

Type to represent a table constraint that would be part of a CREATE TABLE or ALTER TABLE statement. For instances, the UNIQUE constraint in

CREATE TABLE FOO
 ( id integer
 , UNIQUE id
 )

TableConstraint provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

uniqueConstraint :: NonEmpty ColumnName -> TableConstraint

Constructs a TableConstraint will create a UNIQUE constraint on the given columns.

Since: 1.0.0.0

foreignKeyConstraint

Arguments

:: NonEmpty ColumnName
The names of the columns in the source table that form the foreign key

-> Qualified TableName
The table of the table that the foreign key references

-> NonEmpty ColumnName
The names of the columns in the foreign table that the foreign key references

-> Maybe ForeignKeyUpdateActionExpr
An optional ON UPDATE foreign key action to perform

-> Maybe ForeignKeyDeleteActionExpr
An optional ON DELETE foreign key action to perform

-> TableConstraint

Constructs a TableConstraint that represent a FOREIGN KEY constraint

Since: 1.0.0.0

data ForeignKeyActionExpr

Type to represent a foreign key action on a FOREIGN KEY constraint. E.G. the CASCADE in

FOREIGN KEY (foo_id) REFERENCES foo (id) ON DELETE CASCADE

ForeignKeyActionExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

restrictExpr :: ForeignKeyActionExpr

The foreign key action RESTRICT.

Since: 1.0.0.0

cascadeExpr :: ForeignKeyActionExpr

The foreign key action CASCADE.

Since: 1.0.0.0

setNullExpr :: ForeignKeyActionExpr

The foreign key action SET NULL.

Since: 1.0.0.0

setDefaultExpr :: ForeignKeyActionExpr

The foreign key action SET DEFAULT.

Since: 1.0.0.0

data ForeignKeyDeleteActionExpr

Type to represent an foreign key update action on a FOREIGN KEY constraint. E.G. the ON DELETE RESTRICT in

FOREIGN KEY (foo_id) REFERENCES foo (id) ON DELETE RESTRICT

ForeignKeyDeleteActionExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

foreignKeyDeleteActionExpr :: ForeignKeyActionExpr -> ForeignKeyDeleteActionExpr

Constructs a ForeignKeyActionExpr that use the given ForeignKeyActionExpr in an ON UPDATE clause for a foreign key.

Since: 1.0.0.0

data ForeignKeyUpdateActionExpr

Type to represent an foreign key update action on a FOREIGN KEY constraint. E.G. the ON UPDATE RESTRICT in

FOREIGN KEY (foo_id) REFERENCES foo (id) ON UPDATE RESTRICT

ForeignKeyUpdateActionExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

foreignKeyUpdateActionExpr :: ForeignKeyActionExpr -> ForeignKeyUpdateActionExpr

Constructs a ForeignKeyActionExpr that use the given ForeignKeyActionExpr in an ON UPDATE clause for a foreign key.

Since: 1.0.0.0