Orville.PostgreSQL.Expr.TableDefinition

Description

Since: 1.0.0.0

Documentation

data CreateTableExpr

Type to represent a CREATE TABLE statement. E.G.

CREATE TABLE foo (id integer)

CreateTableExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

createTableExpr

Arguments

:: Qualified TableName
The name to be use for the table

-> [ColumnDefinition]
The columns to include in the table

-> Maybe PrimaryKeyExpr
A primary key expression for the table

-> [TableConstraint]
Any table constraints to include with the table

-> CreateTableExpr

Constructs a CreateTableExpr with the given options.

Since: 1.0.0.0

data PrimaryKeyExpr

Type to represent a the primary key of a table. E.G.

PRIMARY KEY (id)

PrimaryKeyExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

primaryKeyExpr :: NonEmpty ColumnName -> PrimaryKeyExpr

Constructs a PrimaryKeyExpr with the given columns

Since: 1.0.0.0

data AlterTableExpr

Type to represent a ALTER TABLE statement. E.G.

ALTER TABLE foo ADD COLUMN bar integer

AlterTableExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

alterTableExpr :: Qualified TableName -> NonEmpty AlterTableAction -> AlterTableExpr

Constructs an AlterTableExpr with the given alter table actions.

Since: 1.0.0.0

data AlterTableAction

Type to represent an action as part of an ALTER TABLE statement. E.G.

ADD COLUMN bar integer

AlterTableAction provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

addColumn :: ColumnDefinition -> AlterTableAction

Constructs an AlterTableAction that will add the specified column to the table.

Since: 1.0.0.0

dropColumn :: ColumnName -> AlterTableAction

Constructs an AlterTableAction that will drop the specified column from the table.

Since: 1.0.0.0

addConstraint :: TableConstraint -> AlterTableAction

Constructs an AlterTableAction that will add the specified constraint to the table.

Since: 1.0.0.0

dropConstraint :: ConstraintName -> AlterTableAction

Constructs an AlterTableAction that will drop the specified constraint from the table.

Since: 1.0.0.0

alterColumnType

Arguments

:: ColumnName
The name of the column whose type willbe altered

-> DataType
The new type to use for the column

-> Maybe UsingClause
An optional UsingClause to indicate to the database how data from the old type should be converted to the new type.

-> AlterTableAction

Constructs an AlterTableAction that will alter the type of the specified column.

Since: 1.0.0.0

alterColumnSetDefault :: SqlExpression valueExpression => ColumnName -> valueExpression -> AlterTableAction

Constructs an AlterTableAction that will use SET DEFAULT to set the default value of the specified column.

Since: 1.0.0.0

alterColumnDropDefault :: ColumnName -> AlterTableAction

Constructs an AlterTableAction that will use DROP DEFAULT to drop the default value of the specified column.

Since: 1.0.0.0

data UsingClause

Type to represent a USING clause as part of an ALTER COLUMN when changing the type of a column. E.G.

USING id :: integer

UsingClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

usingCast :: ColumnName -> DataType -> UsingClause

Constructs a UsingClause that will cast the column to the specified type.

Since: 1.0.0.0

alterColumnNullability :: ColumnName -> AlterNotNull -> AlterTableAction

Constructs an AlterTableAction that will alter the nullability of the column.

Since: 1.0.0.0

data AlterNotNull

Type to represent an action to alter the nullability of a column. E.G.

SET NOT NULL

AlterNotNull provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

setNotNull :: AlterNotNull

Sets the column to not null via SET NOT NULL

Since: 1.0.0.0

dropNotNull :: AlterNotNull

Sets the column to allow null via DROP NOT NULL

Since: 1.0.0.0

data DropTableExpr

Type to represent a DROP TABLE statement. E.G.

DROP TABLE FOO

DropTableExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

dropTableExpr :: Maybe IfExists -> Qualified TableName -> DropTableExpr

Constructs an DropTableExpr that will drop the specified table.

Since: 1.0.0.0