Orville.PostgreSQL.Expr.TableReferenceList

Description

Since: 1.0.0.0

Documentation

data TableReferenceList

Type to represent the tables references in the FROM clause of a @SELECT statement. E.G. just the

foo

in

FROM foo

TableReferenceList provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

referencesTable :: Qualified TableName -> TableReferenceList

Constructs a TableReferenceList consisting of just the specified table name.

Since: 1.0.0.0