Orville.PostgreSQL.Expr.Time

Description

Since: 1.0.0.0

Documentation

now :: ValueExpression

The value of the current time as returned by the PostgreSQL function now().

Since: 1.0.0.0

makeInterval :: [(IntervalArgument, ValueExpression)] -> ValueExpression

Constructs a ValueExpression whose value in PostgreSQL is the result of calling make_interval with the specified time intervals passed as named arguments.

Since: 1.0.0.0

data IntervalArgument

Type to represent the name of a time interval argument to the PostgreSQL make_interval function. E.G.

years

IntervalArgument provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

years :: IntervalArgument

The years argument to make_interval.

Since: 1.0.0.0

months :: IntervalArgument

The months argument to make_interval.

Since: 1.0.0.0

weeks :: IntervalArgument

The weeks argument to make_interval.

Since: 1.0.0.0

days :: IntervalArgument

The days argument to make_interval.

Since: 1.0.0.0

hours :: IntervalArgument

The hours argument to make_interval.

Since: 1.0.0.0

minutes :: IntervalArgument

The mins argument to make_interval.

Since: 1.0.0.0

seconds :: IntervalArgument

The secs argument to make_interval.

Since: 1.0.0.0