Orville.PostgreSQL.Expr.Transaction

Description

Since: 1.0.0.0

Documentation

data BeginTransactionExpr

Type to represent the name of a begin transaction statement. E.G.

BEGIN TRANSACTION

BeginTransactionExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

beginTransaction :: Maybe TransactionMode -> BeginTransactionExpr

Constructs a BeginTransactionExpr that will begin a transaction using the specified mode, if any.

Since: 1.0.0.0

data TransactionMode

Type to represent the a transaction mode. E.G.

ISOLATION LEVEL SERIALIZABLE

TransactionMode provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

readWrite :: TransactionMode

The READ WRITE transaction mode.

Since: 1.0.0.0

readOnly :: TransactionMode

The READ ONLY transaction mode.

Since: 1.0.0.0

deferrable :: TransactionMode

The DEFERRABLE transaction mode.

Since: 1.0.0.0

notDeferrable :: TransactionMode

The NOT DEFERRABLE transaction mode.

Since: 1.0.0.0

isolationLevel :: IsolationLevel -> TransactionMode

An ISOLATION LEVEL transaction mode with the given IsolationLevel.

Since: 1.0.0.0

data IsolationLevel

Type to represent the a transaction isolation level. E.G.

SERIALIZABLE

IsolationLevel provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

serializable :: IsolationLevel

The SERIALIZABLE isolation level.

Since: 1.0.0.0

repeatableRead :: IsolationLevel

The REPEATABLE READ isolation level.

Since: 1.0.0.0

readCommitted :: IsolationLevel

The READ COMMITTED isolation level.

Since: 1.0.0.0

readUncommitted :: IsolationLevel

The READ UNCOMMITTED isolation level.

Since: 1.0.0.0

data CommitExpr

Type to represent the transaction commit statement. E.G.

COMMIT

CommitExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

commit :: CommitExpr

A COMMIT transaction statement

Since: 1.0.0.0

data RollbackExpr

Type to represent the transaction rollback statement. E.G.

ROLLBACK

RollbackExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

rollback :: RollbackExpr

A ROLLBACK transaction statement

Since: 1.0.0.0

rollbackTo :: SavepointName -> RollbackExpr

A ROLLBACK TO transaction statement that will rollback to the specified savepoint.

Since: 1.0.0.0

data SavepointExpr

Type to represent the transaction savepoint statement. E.G.

SAVEPOINT foo

SavepointExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

savepoint :: SavepointName -> SavepointExpr

A SAVEPOINT statement that will create a savepoint with the given name.

Since: 1.0.0.0

data ReleaseSavepointExpr

Type to represent the transaction release savepoint statement. E.G.

RELEASE SAVEPOINT foo

ReleaseSavepointExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

releaseSavepoint :: SavepointName -> ReleaseSavepointExpr

A RELEASE SAVEPOINT statement that will release the specified savepoint.

Since: 1.0.0.0