Orville.PostgreSQL.Expr.Update

Description

Since: 1.0.0.0

Documentation

data UpdateExpr

Type to represent the transaction a SQL UPDATE statement. E.G.

UPDATE foo
SET id = 1
WHERE id <> 1

UpdateExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

updateExpr

Arguments

:: Qualified TableName
The name of the table to be updated

-> SetClauseList
The updates to be made to the table

-> Maybe WhereClause
An optional where clause to limit the rows updated

-> Maybe ReturningExpr
An optional returning clause to return data from the updated rows

-> UpdateExpr

Constructs an UpdateExpr with the given options.

Since: 1.0.0.0

data SetClauseList

Type to represent the list of updates to be made in a UPDATE statament. E.G.

foo = 1,
bar = 2

SetClauseList provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

setClauseList :: NonEmpty SetClause -> SetClauseList

Constructs a SetClauseList with the specified set clauses

Since: 1.0.0.0

data SetClause

Type to represent a single updates to be made in a UPDATE statament. E.G.

foo = 1

SetClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

setColumn :: ColumnName -> SqlValue -> SetClause

Constructs a SetClause that will set the specified column to the specified value.

Since: 1.0.0.0