Orville.PostgreSQL.Expr.WhereClause

Description

Since: 1.0.0.0

Documentation

data WhereClause

Type to represent a WHERE clause restriction an a SELECT, UPDATE or DELETE statement. E.G.

WHERE (foo > 10)

WhereClause provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

whereClause :: BooleanExpr -> WhereClause

Constructs a WHERE clause from the given BooleanExpr. E.G.

WHERE <boolean expr>

Since: 1.0.0.0

data BooleanExpr

Type to represent a SQL value expression that evaluates to a boolean and thereforce can used with boolean logic functions. E.G.

foo > 10

BooleanExpr provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

literalBooleanExpr :: Bool -> BooleanExpr

Constructs a BooleanExpr whose value the SQL literal TRUE or FALSE depending on the argument given.

Since: 1.0.0.0

andExpr :: BooleanExpr -> BooleanExpr -> BooleanExpr

The SQL AND operator. The arguments will be surrounded with parentheses to ensure that the associativity of expression in the resulting SQL matches the associtivity implied by this Haskell function.

Since: 1.0.0.0

(.&&) :: BooleanExpr -> BooleanExpr -> BooleanExpr infixr 8

The SQL AND operator (alias for andExpr)

Since: 1.0.0.0

orExpr :: BooleanExpr -> BooleanExpr -> BooleanExpr

The SQL OR operator. The arguments will be surrounded with parentheses to ensure that the associativity of expression in the resulting SQL matches the associtivity implied by this Haskell function.

Since: 1.0.0.0

(.||) :: BooleanExpr -> BooleanExpr -> BooleanExpr infixr 8

The SQL OR operator (alias for orExpr)

Since: 1.0.0.0

parenthesized :: BooleanExpr -> BooleanExpr

Surrounds the given BooleanExpr with parentheses.

Since: 1.0.0.0

equals :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL = operator.

Since: 1.0.0.0

notEquals :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL <> operator.

Since: 1.0.0.0

greaterThan :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL > operator.

Since: 1.0.0.0

lessThan :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL < operator.

Since: 1.0.0.0

greaterThanOrEqualTo :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL >= operator.

Since: 1.0.0.0

lessThanOrEqualTo :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL <= operator.

Since: 1.0.0.0

like :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL LIKE operator.

Since: 1.0.0.0

likeInsensitive :: ValueExpression -> ValueExpression -> BooleanExpr

The SQL ILIKE operator.

Since: 1.0.0.0

isNull :: ValueExpression -> BooleanExpr

The SQL IS NULL condition.

Since: 1.0.0.0

isNotNull :: ValueExpression -> BooleanExpr

The SQL IS NOT NULL condition.

Since: 1.0.0.0

valueIn :: ValueExpression -> NonEmpty ValueExpression -> BooleanExpr

The SQL IN operator. The result will be TRUE if the given value appears in the list of values given.

Since: 1.0.0.0

valueNotIn :: ValueExpression -> NonEmpty ValueExpression -> BooleanExpr

The SQL IN operator. The result will be TRUE if the given value does not appear in the list of values given.

Since: 1.0.0.0

tupleIn :: NonEmpty ValueExpression -> NonEmpty (NonEmpty ValueExpression) -> BooleanExpr

The SQL IN operator, like valueIn, but for when you want to construct a tuple in SQL and check if it is in a list of tuples. It is up to the caller to ensure that all the tuples given have the same arity.

Since: 1.0.0.0

tupleNotIn :: NonEmpty ValueExpression -> NonEmpty (NonEmpty ValueExpression) -> BooleanExpr

The SQL NOT IN operator, like valueNotIn, but for when you want to construct a tuple in SQL and check if it is not in a list of tuples. It is up to the caller to ensure that all the tuples given have the same arity.

Since: 1.0.0.0

data InValuePredicate

Type to represent the right hand side of an IN or NOT IN expression. E.G.

(10,12,13)

InValuePredicate provides a SqlExpression instance. See unsafeSqlExpression for how to construct a value with your own custom SQL.

Since: 1.0.0.0

inPredicate :: ValueExpression -> InValuePredicate -> BooleanExpr

Lowel lever access to the the SQL IN operator. This takes any ValueExpression and InValuePredicate. It is up to the caller to ensure the expressions given makes sense together.

Since: 1.0.0.0

notInPredicate :: ValueExpression -> InValuePredicate -> BooleanExpr

Lowel lever access to the the SQL NOT IN operator. This takes any ValueExpression and InValuePredicate. It is up to the caller to ensure the expressions given makes sense together.

Since: 1.0.0.0

inValueList :: NonEmpty ValueExpression -> InValuePredicate

Constructs an InValuePredicate from the given list of ValueExpression

Since: 1.0.0.0