Orville.PostgreSQL.Marshall

Description

You can import Orville.PostgreSQL.Marshall to get access to all the functions related to marshalling Haskell values to and from their SQL representations. This includes a number of lowel-level items not exported by Orville.PostgreSQL that give you more control (and therefore responsibility) over how the marshalling is performed.

Since: 1.0.0.0

Documentation

module Orville.PostgreSQL.Marshall.SqlMarshaller

module Orville.PostgreSQL.Marshall.FieldDefinition

module Orville.PostgreSQL.Marshall.DefaultValue

module Orville.PostgreSQL.Marshall.SyntheticField

module Orville.PostgreSQL.Marshall.MarshallError

module Orville.PostgreSQL.Marshall.SqlType