Orville.PostgreSQL.Marshall.MarshallError

Description

Since: 1.0.0.0

Documentation

data MarshallError

A MarshallError may be returned from marshallResultFromSql when a row being decoded from the database doesn't meet the expectations of the SqlMarshaller that is decoding it.

Since: 1.0.0.0

Constructors

MarshallError

Fields

marshallErrorDetailLevel :: ErrorDetailLevel

The level of details that will be used to render this error as a message if show is called

marshallErrorRowIdentifier :: [(ByteString, SqlValue)]

The identifier of the row that caused the error. This is a list of pairs of column name and value in their raw form from the database to avoid further possible decoding errors when reading the values

marshallErrorDetails :: MarshallErrorDetails

The detailed information about the error that occurred during decoding

renderMarshallError :: ErrorDetailLevel -> MarshallError -> String

Renders a MarshallError to a string using the specified ErrorDetailLevel.

This ingores any ErrorDetailLevel that was captured by default from the Orville context and uses the specified level of detail instead.

You may want to use this function to render certain with a higher level of detail that you consider safe for (for example) you application logs while using a lower default error detail level to be used with the Show instance of MarhallError in case an exception is handled in a more visible section of code that return information more publicly (e.g. a request handler for a public endpoint).

Since: 1.0.0.0

data MarshallErrorDetails

A MarshallErrorDetails may be returned from marshallFromSql if the result set being decoded from the database doesn't meet the expectations of the SqlMarshaller that is decoding it.

Since: 1.0.0.0

Constructors

DecodingError DecodingErrorDetails
Indicates that a one ore more values in a columns could not be decoded, either individually or as a group

MissingColumnError MissingColumnErrorDetails
Indicates that an expected column was not found in the result set

renderMarshallErrorDetails :: ErrorDetailLevel -> MarshallErrorDetails -> String

Renders a MarshallErrorDetails to a String with a specified ErrorDetailLevel.

Since: 1.0.0.0

data DecodingErrorDetails

Details about an error that occurred while decoding values found a SQL result set.

Since: 1.0.0.0

Constructors

DecodingErrorDetails

Fields

decodingErrorValues :: [(ByteString, SqlValue)]

decodingErrorMessage :: String

renderDecodingErrorDetails :: ErrorDetailLevel -> DecodingErrorDetails -> String

Renders a 'DecodingErrorDetails to a String with a specified ErrorDetailLevel.

Since: 1.0.0.0

data MissingColumnErrorDetails

Details about an column that was found to be missing in a SQL result set during decoding.

Since: 1.0.0.0

Constructors

MissingColumnErrorDetails

Fields

missingColumnName :: ByteString

actualColumnNames :: Set ByteString

renderMissingColumnErrorDetails :: ErrorDetailLevel -> MissingColumnErrorDetails -> String

Renders a MissingColumnErrorDetails to a String with a specified ErrorDetailLevel.

Since: 1.0.0.0