Orville.PostgreSQL.Marshall.SyntheticField

Description

Since: 1.0.0.0

Documentation

data SyntheticField a

A SyntheticField can be used to evaluate a SQL expression based on the columns of a table when records are selected from the database. Synthetic fields are inherently read-only.

Since: 1.0.0.0

syntheticFieldExpression :: SyntheticField a -> ValueExpression

Returns the SQL expression that should be in with select statements to calculated the sythetic field.

Since: 1.0.0.0

syntheticFieldAlias :: SyntheticField a -> FieldName

Returns the alias that should be used in select statements to name the the synthetic field.

Since: 1.0.0.0

syntheticFieldValueFromSqlValue :: SyntheticField a -> SqlValue -> Either String a

Decodes a calculated value selected from the database to its expected Haskell type. Returns a Left with an error message if the decoding fails.

Since: 1.0.0.0

syntheticField

Arguments

:: ValueExpression
The SQL expression to be selected

-> String
The alias to be used to name the calculation in SQL experios

-> (SqlValue -> Either String a)
A function to decode the expression result from a SqlValue

-> SyntheticField a

Constructs a SyntheticField that will select a SQL expression using the given alias.

Since: 1.0.0.0

nullableSyntheticField :: SyntheticField a -> SyntheticField (Maybe a)

Modifies a SyntheticField to allow it to decode NULL values.

Since: 1.0.0.0

prefixSyntheticField :: String -> SyntheticField a -> SyntheticField a

Adds a prefix, followed by an underscore, to the alias used to name the synthetic field.

Since: 1.0.0.0