Orville.PostgreSQL.OrvilleState

Description

Since: 1.0.0.0

Documentation

data OrvilleState

OrvilleState is used to manange opening connections to the database, transactions, etc. newOrvilleState should be used to create an appopriate initial state for your monad's context.

Since: 1.0.0.0

newOrvilleState :: ErrorDetailLevel -> Pool Connection -> OrvilleState

Creates a appropriate initial OrvilleState that will use the connection pool given to initiate connections to the database.

Since: 1.0.0.0

resetOrvilleState :: OrvilleState -> OrvilleState

Creates a new initial OrvilleState using the connection pool from the provide state. You might need to use this if you are spawning one Orville monad from another and they should not share the same connection and transaction state.

Since: 1.0.0.0

orvilleConnectionPool :: OrvilleState -> Pool Connection

Get the connection pool being used for the OrvilleState

Since: 1.0.0.0

orvilleErrorDetailLevel :: OrvilleState -> ErrorDetailLevel

The ErrorDetailLevel controls how much information Orville includes in error messages it generates when data cannot be decoded rows in the database.

Since: 1.0.0.0

orvilleTransactionCallback :: OrvilleState -> TransactionEvent -> IO ()

Orville will call the transaction callback any time a transaction event occurrs. You can register a callback with addTransactionCallBack.

Since: 1.0.0.0

orvilleSqlCommenterAttributes :: OrvilleState -> Maybe SqlCommenterAttributes

The SqlCommenter attributes that Orville will include with queries. These can be modified with addSqlCommenterAttributes. See https://google.github.io/sqlcommenter/.

Since: 1.0.0.0

addTransactionCallback :: (TransactionEvent -> IO ()) -> OrvilleState -> OrvilleState

Registers a callback to be invoked during transactions.

The callback given will be called after the SQL statement corresponding to the given event has finished executing. Callbacks will be called in the order the are added.

Note: There is no specialized error handling for these callbacks. This means that if a callback raises an exception no further callbacks will be called and the exception will propagate up until it caught elsewhere. In particular, if an exception is raised by a callback upon opening the transaction it will cause the transaction to be rolled-back the same as any other exception that might happen during the transaction. In general, we recommend only using callbacks that either raise no exceptions or can handle their own exceptions cleanly.

Since: 1.0.0.0

data TransactionEvent

Describes an event in the lifecycle of a database transaction. You can use addTransactionCallBack to register a callback to respond to these events. The callback will be called after the even in question has been succesfully executed.

Since: 1.0.0.0

Constructors

BeginTransaction
Indicates a new transaction has been started

NewSavepoint Savepoint
Indicates that a new savepoint has been saved within a transaction

ReleaseSavepoint Savepoint
Indicates that a previous savepoint has been released. It can no longer be rolled back to.

RollbackToSavepoint Savepoint
Indicates that rollbac was performed to a prior savepoint.

Note: It is possible to rollback to a savepoint prior to the most recent one without releasing or rolling back to intermediate savepoints. Doing so destroys any savepoints created after given savepoint. Although Orville currently always matches NewSavepoint with either ReleaseSavepoint or RollbackToSavepoint, it is recommended that you do not rely on this behavior.

CommitTransaction
Indicates that the transaction has been committed.

RollbackTransaction
Indicates that the transaction has been rolled back.

data Savepoint

An internal Orville identifier for a savepoint in a PostgreSQL transaction.

Since: 1.0.0.0

savepointNestingLevel :: Savepoint -> Int

Indicates how many levels of nested savepoints the given Savepoint identifier represents.

Since: 1.0.0.0

initialSavepoint :: Savepoint

The initial identifier Orville uses to track the first savepoint within a transaction.

Since: 1.0.0.0

nextSavepoint :: Savepoint -> Savepoint

Determines identifier for the next savepoint in a transaction after the given saveponit>

Since: 1.0.0.0

orvilleSqlExecutionCallback :: OrvilleState -> forall a. QueryType -> RawSql -> IO a -> IO a

The callback Orville will call whenever it wants to run SQL. You can register a callback using addSqlExecutionCallback.

Since: 1.0.0.0

addSqlExecutionCallback :: (forall a. QueryType -> RawSql -> IO a -> IO a) -> OrvilleState -> OrvilleState

Adds a callback to be called when an Orville operation executes a SQL statement. The callback is given the IO action that will perform the query execution and must call that action for the query to be run. In particular, you can use this to time query and log any that are slow.

Calls to any previously added callbacks will also be execute as part of the IO action passed to the new callback. Thus the newly added callback happens "around" the previously added callback.

There is no special exception handling done for these callbacks beyond what they implement themelves. Any callbacks should allow for the possibility that the IO action they are given may raise an exception.

Since: 1.0.0.0

orvilleBeginTransactionExpr :: OrvilleState -> BeginTransactionExpr

The SQL expression that Orville will use to begin a transaction. You can set this via setBeginTransactionExpr to have fine grained control over the transaction parameters, such as isolation level.

Since: 1.0.0.0

setBeginTransactionExpr :: BeginTransactionExpr -> OrvilleState -> OrvilleState

Sets the SQL expression that Orville will use to begin transactions. You can control the transaction isolation level by building your own BeginTransactionExpr with the desired isolation level.

Since: 1.0.0.0

setSqlCommenterAttributes :: SqlCommenterAttributes -> OrvilleState -> OrvilleState

Sets the SqlCommenterAttributes that Orville will then add to any following statement executions.

Since: 1.0.0.0

addSqlCommenterAttributes :: SqlCommenterAttributes -> OrvilleState -> OrvilleState

Adds the SqlCommenterAttributes to the already existing that Orville will then add to any following statement executions.

Since: 1.0.0.0