Orville.PostgreSQL.PgCatalog

Description

Since: 1.0.0.0

Documentation

data PgSequence

The Haskell representation of data read from the pg_catalog.pg_sequence table. Rows in this table sequences in PostgreSQL.

Since: 1.0.0.0

Constructors

PgSequence

Fields

pgSequenceClassOid :: Oid

The PostgreSQL oid of the pg_class for this sequence

pgSequenceTypeOid :: Oid

The PostgreSQL oid of the data type of the sequence. References pg_type.oid

pgSequenceStart :: Int64

The start value of the sequence

pgSequenceIncrement :: Int64

The increment value of the sequence

pgSequenceMax :: Int64

The max value of the sequence

pgSequenceMin :: Int64

The max value of the sequence

pgSequenceCache :: Int64

The cache size of the sequence

pgSequenceCycle :: Bool

Wether the sequence cycles

pgSequenceTable :: TableDefinition (HasKey Oid) PgSequence PgSequence

An Orville TableDefinition for querying the pg_catalog.pg_sequence table

Since: 1.0.0.0

sequencePgClassOidField :: FieldDefinition NotNull Oid

The seqrelid column of the pg_cataglog.pg_sequence table

Since: 1.0.0.0

data PgNamespace

The Haskell representation of data read from the pg_catalog.pg_namespace table. Namespaces in pg_catalog correspond to "schema" concept in database organization.

Since: 1.0.0.0

Constructors

PgNamespace

Fields

pgNamespaceOid :: Oid

The PostgreSQL oid for the namespace. This is referenced from other tables, such as pg_class.

pgNamespaceName :: NamespaceName

The name of the namespace.

data NamespaceName

A Haskell type for the name of a namespace

Since: 1.0.0.0

namespaceNameToString :: NamespaceName -> String

Convert a 'NamespaceName to a plain String

Since: 1.0.0.0

pgNamespaceTable :: TableDefinition (HasKey Oid) PgNamespace PgNamespace

An Orville TableDefinition for querying the pg_catalog.pg_namespace table

Since: 1.0.0.0

namespaceNameField :: FieldDefinition NotNull NamespaceName

The nspname column of the pg_catalog.pg_namespace table

Since: 1.0.0.0

data PgIndex

The Haskell representation of data read from the pg_catalog.pg_index tale. Rows in this table contain extended information about indices. Information about indices is also contained in the pg_catalog.pg_class table as well.

Since: 1.0.0.0

Constructors

PgIndex

Fields

pgIndexPgClassOid :: Oid

The PostgreSQL oid of the pg_class entry for this index.

pgIndexRelationOid :: Oid

The PostgreSQL oid of the pg_class entry for the table that this index is for.

pgIndexAttributeNumbers :: [AttributeNumber]

An array of attribute numbers references the columns the table that are included in the index. An attribute number of 0 indicates an expression over the table's columns rather than just a reference to a column.

In PostgreSQL 11+ this includes both key columns and non-key included columns. Orville is currently not aware of this distinction, however.

pgIndexIsUnique :: Bool

Indicates whether this is a unique index

pgIndexIsPrimary :: Bool

Indicates whether this is the primary key index for the table

pgIndexIsLive :: Bool

When False, indicates that this index is in the process of being dropped and should be ignored

pgIndexTable :: TableDefinition NoKey PgIndex PgIndex

An Orville TableDefinition for querying the pg_catalog.pg_index table

Since: 1.0.0.0

indexRelationOidField :: FieldDefinition NotNull Oid

The indrelid column of the pg_index table

Since: 1.0.0.0

indexIsLiveField :: FieldDefinition NotNull Bool

The indislive column of the pg_index table

Since: 1.0.0.0

data PgConstraint

The Haskell representation of data read from the pg_catalog.pg_constraint tale. Rows in this table correspond to check, primary key, unique, foreign key and exclusion constraints on tables.

Since: 1.0.0.0

Constructors

PgConstraint

Fields

pgConstraintOid :: Oid

The PostgreSQL oid for the constraint

pgConstraintName :: ConstraintName

The constraint name (which may not be unique)

pgConstraintNamespaceOid :: Oid

The oid of the namespace that contains the constraint

pgConstraintType :: ConstraintType

The type of constraint

pgConstraintRelationOid :: Oid

The PostgreSQL oid of the table that the constraint is on (or 0 if not a table constraint)

pgConstraintIndexOid :: Oid

The PostgreSQL oid ef the index supporting this constraint, if it's a unique, primary key, foreign key or exclusion constraint. Otherwise 0.

pgConstraintKey :: Maybe [AttributeNumber]

For table constraints, the attribute numbers of the constrained columns. These correspond to the pgAttributeNumber field of PgAttribute.

pgConstraintForeignRelationOid :: Oid

For foreign key constraints, the PostgreSQL oid of the table the foreign key references

pgConstraintForeignKey :: Maybe [AttributeNumber]

For foreignkey constraints, the attribute numbers of the referenced columns. These correspond to the pgAttributeNumber field of PgAttribute.

pgConstraintForeignKeyOnUpdateType :: Maybe ForeignKeyAction

For foreignkey constraints, the on update action type

pgConstraintForeignKeyOnDeleteType :: Maybe ForeignKeyAction

For foreignkey constraints, the on delete action type

data ConstraintType

The type of constraint that a PgConstraint represents, as described at https://www.postgresql.org/docs/13/catalog-pg-constraint.html

Since: 1.0.0.0

Constructors

CheckConstraint

ForeignKeyConstraint

PrimaryKeyConstraint

UniqueConstraint

ConstraintTrigger

ExclusionConstraint

data ConstraintName

A Haskell type for the name of the constraint represented by a PgConstraint

Since: 1.0.0.0

constraintNameToString :: ConstraintName -> String

Converts an ConstraintName to a plain old string

Since: 1.0.0.0

pgConstraintTable :: TableDefinition (HasKey Oid) PgConstraint PgConstraint

An Orville TableDefinition for querying the pg_catalog.pg_constraint table

Since: 1.0.0.0

constraintRelationOidField :: FieldDefinition NotNull Oid

The conrelid column of the pg_constraint table

Since: 1.0.0.0

data PgClass

The Haskell representation of data read from the pg_catalog.pg_class table. Rows in this table correspond to tables, indexes, sequences, views, materialized views, composite types and TOAST tables.

Since: 1.0.0.0

Constructors

PgClass

Fields

pgClassOid :: Oid

The PostgreSQL oid for the relation

pgClassNamespaceOid :: Oid

The PostgreSQL oid of the namespace that the relation belongs to. References pg_namespace.oid.

pgClassRelationName :: RelationName

The name of relation

pgClassRelationKind :: RelationKind

The kind of relation (table, view, etc)

data RelationName

A Haskell type for the name of the relation represented by a PgClass

Since: 1.0.0.0

relationNameToString :: RelationName -> String

Convert a RelationName to a plain String

Since: 1.0.0.0

data RelationKind

The kind of relation represented by a PgClass, as described at https://www.postgresql.org/docs/13/catalog-pg-class.html.

Since: 1.0.0.0

Constructors

OrdinaryTable

Index

Sequence

ToastTable

View

MaterializedView

CompositeType

ForeignTable

PartitionedTable

PartitionedIndex

pgClassTable :: TableDefinition (HasKey Oid) PgClass PgClass

An Orville TableDefinition for querying the pg_catalog.pg_class table

Since: 1.0.0.0

relationNameField :: FieldDefinition NotNull RelationName

The relname column of the pg_catalog.pg_class table

Since: 1.0.0.0

namespaceOidField :: FieldDefinition NotNull Oid

The relnamespace column of the pg_catalog.pg_class table

Since: 1.0.0.0

relationKindField :: FieldDefinition NotNull RelationKind

The relkind column of the pg_catalog.pg_class table

Since: 1.0.0.0

data PgAttributeDefault

The Haskell representation of data read from the pg_catalog.pg_attrdef table.

Since: 1.0.0.0

Constructors

PgAttributeDefault

Fields

pgAttributeDefaultOid :: Oid

The PostgreSQL oid for the default value

pgAttributeDefaultRelationOid :: Oid

The PostgreSQL oid for the relation that this attribute belongs to. References pg_class.oid

pgAttributeDefaultAttributeNumber :: AttributeNumber

The PostgreSQL attribute number for the column that this default belongs to. References pg_attribute.attnum.

pgAttributeDefaultExpression :: Text

The PostgreSQL default value expression, as decompiled from the adbin column using the PostgreSQL pg_get_expr function.

pgAttributeDefaultTable :: TableDefinition NoKey PgAttributeDefault PgAttributeDefault

An Orville TableDefinition for querying the pg_catalog.pg_attrdef table

Since: 1.0.0.0

attributeDefaultRelationOidField :: FieldDefinition NotNull Oid

The adrelid column of the pg_catalog.pg_attrdef table

Since: 1.0.0.0

data PgAttribute

The Haskell representation of data read from the pg_catalog.pg_attribute table. Rows in this table correspond to table columns, but also for attributes of other items from the pg_class table.

See also PgClass.

Since: 1.0.0.0

Constructors

PgAttribute

Fields

pgAttributeRelationOid :: Oid

The PostgreSQL oid for the relation that this attribute belongs to. References pg_class.oid

pgAttributeName :: AttributeName

The name of attribute

pgAttributeNumber :: AttributeNumber

The PostgreSQL number of attribute

pgAttributeTypeOid :: Oid

The PostgreSQL oid for the type of this attribute. References pg_type.oid

pgAttributeLength :: Int16

The length of this attributes type (a copy of pg_type.typlen). Note that this is _NOT_ the maximum length of a varchar column!

pgAttributeTypeModifier :: Int32

Type-specific data supplied at creation time, such as the maximum length of a varchar column

pgAttributeIsDropped :: Bool

Indicates whether the column has been dropped and is not longer valid

pgAttributeIsNotNull :: Bool

Indicates whether the column has a not-null constraint

pgAttributeMaxLength :: PgAttribute -> Maybe Int32

Returns the maximum length for an attribute with a variable length type, or Nothing if the length if the type is not variable.

Since: 1.0.0.0

data AttributeName

A Haskell type for the name of the attribute represented by a PgAttribute

Since: 1.0.0.0

attributeNameToString :: AttributeName -> String

Converts an AttributeName to a plain old string

Since: 1.0.0.0

data AttributeNumber

A Haskell type for the number of the attribute represented by a PgAttribute

Since: 1.0.0.0

attributeNumberToInt16 :: AttributeNumber -> Int16

Converts an AttributeNumber to an integer

attributeNumberFromInt16 :: Int16 -> AttributeNumber

Converts an integer to an AttributeNumber

attributeNumberTextBuilder :: AttributeNumber -> Builder

Encodes an AttributeNumber to lazy text as a builder

Since: 1.0.0.0

attributeNumberParser :: Parser AttributeNumber

Attoparsec parser for AttributeNumber

Since: 1.0.0.0

isOrdinaryColumn :: PgAttribute -> Bool

Determines whether the attribute represents a system column by inspecting the attribute's AttributeNumber. Ordinary columns have attribute numbers starting at 1.

Since: 1.0.0.0

pgAttributeTable :: TableDefinition NoKey PgAttribute PgAttribute

An Orville TableDefinition for querying the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeRelationOidField :: FieldDefinition NotNull Oid

The attrelid column of the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeNameField :: FieldDefinition NotNull AttributeName

The attname column of the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeTypeOidField :: FieldDefinition NotNull Oid

The atttypid column of the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeLengthField :: FieldDefinition NotNull Int16

The attlen column of the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeIsDroppedField :: FieldDefinition NotNull Bool

The attisdropped column of the pg_catalog.pg_attribute table

Since: 1.0.0.0

attributeNumberTypeField :: String -> FieldDefinition NotNull AttributeNumber

Builds a FieldDefinition for a field with type AttributeNumber

Since: 1.0.0.0

oidField :: FieldDefinition NotNull Oid

The oid field found on many (but not all!) pg_catalog tables

Since: 1.0.0.0

oidTypeField :: String -> FieldDefinition NotNull Oid

Builds a FieldDefinition with the given column name that stores an oid value.

Since: 1.0.0.0

data DatabaseDescription

A description of selected items from a single PostgreSQL database. describeDatabaseRelations can be used to load the descriptions of request items.

Since: 1.0.0.0

Constructors

DatabaseDescription

Fields

databaseRelations :: Map (NamespaceName, RelationName) RelationDescription

data RelationDescription

A description of a particular relation in the PostgreSQL database, including the attributes of the relation.

Since: 1.0.0.0

Constructors

RelationDescription

Fields

relationRecord :: PgClass

relationAttributes :: Map AttributeName PgAttribute

relationAttributeDefaults :: Map AttributeNumber PgAttributeDefault

relationConstraints :: [ConstraintDescription]

relationIndexes :: [IndexDescription]

relationSequence :: Maybe PgSequence

data ConstraintDescription

A description of a particular constraint in the PostgreSQL database, including the attributes and relations that it references.

Since: 1.0.0.0

Constructors

ConstraintDescription

Fields

constraintRecord :: PgConstraint

constraintKey :: Maybe [PgAttribute]

constraintForeignRelation :: Maybe ForeignRelationDescription

constraintForeignKey :: Maybe [PgAttribute]

data ForeignRelationDescription

A description of a relation in the PostgreSQL database that is referenced by a foreign key constraint, including the namespace that the relation belongs to.

Since: 1.0.0.0

Constructors

ForeignRelationDescription

Fields

foreignRelationClass :: PgClass

foreignRelationNamespace :: PgNamespace

data IndexDescription

A description of an index in the PostgreSQL database, including the names of the attributes included in the index and the PgClass record of the index itself (NOT the PgClass of the table that the index is for).

Since: 1.0.0.0

Constructors

IndexDescription

Fields

indexRecord :: PgIndex

indexPgClass :: PgClass

indexMembers :: [IndexMember]

data IndexMember

A description of an index member in the PostgreSQL database. If they member is a simple attribute, the PgAttribute for that is provided. If it is an index over an expression, no further description is currently provided.

Since: 1.0.0.0

Constructors

IndexAttribute PgAttribute

IndexExpression

lookupRelation :: (NamespaceName, RelationName) -> DatabaseDescription -> Maybe RelationDescription

Lookup a relation by its qualified name in the pg_catalog schema.

Since: 1.0.0.0

lookupRelationOfKind :: RelationKind -> (NamespaceName, RelationName) -> DatabaseDescription -> Maybe RelationDescription

Lookup a relation by its qualified name in the pg_catalog schema. If the relation is not of the expected kind, Nothing is returned.

Since: 1.0.0.0

lookupAttribute :: AttributeName -> RelationDescription -> Maybe PgAttribute

Find an attribute by name from the RelationDescription

Since: 1.0.0.0

lookupAttributeDefault :: PgAttribute -> RelationDescription -> Maybe PgAttributeDefault

Find an attribute default from the RelationDescription

Since: 1.0.0.0

describeDatabaseRelations :: MonadOrville m => [(NamespaceName, RelationName)] -> m DatabaseDescription

Describes the requested relations in the current database. If any of the relations do not exist, they will not have an entry in the returned description.

Each RelationDescription will contain all the attributes that currently exist for that relation, according to the pg_catalog tables.

Since: 1.0.0.0