Orville.PostgreSQL.Raw.Connection

Description

Since: 1.0.0.0

Orville definitions

data Connection

An Orville handle for a LibPQ connection.

Since: 1.0.0.0

createConnectionPool

Arguments

:: NoticeReporting
Whether or not notice reporting from LibPQ should be enabled

-> Int
Number of stripes in the connection pool

-> NominalDiffTime
Linger time before closing an idle connection

-> Int
Max number of connections to allocate per stripe

-> ByteString
A PostgreSQL connection string

-> IO (Pool Connection)

createConnectionPool allocates a pool of connections to a PostgreSQL server.

Since: 1.0.0.0

data NoticeReporting

An option for createConnectionPool than indicates whether the LibPQ should print notice reports for warnings to the console

Since: 1.0.0.0

Constructors

EnableNoticeReporting

DisableNoticeReporting

executeRaw :: Connection -> ByteString -> [Maybe PgTextFormatValue] -> IO Result

executeRaw runs a given SQL statement returning the raw underlying result.

All handling of stepping through the result set is left to the caller. This potentially leaves connections open much longer than one would expect if all of the results are not iterated through immediately *and* the data copied. Use with caution.

Since: 1.0.0.0

quoteStringLiteral :: Connection -> ByteString -> IO Builder

Escapes and quotes a string for use as a literal within a SQL command that will be execute on the given connection. This uses the PQescapeStringConn function from libpq, which takes the character encoding of the connection into account. Not that while PQescapeStringConn does not surround the literal with quotes, this function does for the sake of symmetry with quoteIdentifier.

This function returns a Buider so that the result can be included in a builder being constructed for the surrounding SQL command without making an additional copy of the Bytestring returned by LibPQ for the sake of adding the surrounding quotes.

Since: 1.0.0.0

quoteIdentifier :: Connection -> ByteString -> IO Builder

Escapes and quotes a string for use as an identifier within a SQL command that will be execute on the given connection. This uses the PQescapeIdentifier function from libpq, which takes the character encoding of the connection into account and also applies the quotes.

Although this function does not need to copy the ByteString returned by LibPQ to add the quotes (since LibPQ already added them), it returns a Builder nonetheless to maintain symmetry with quoteStringLiteral.

Since: 1.0.0.0

data ConnectionUsedAfterCloseError

Orville throws as ConnectionUsedAfterCloseError if it attempts to use a Connection value after it has already been closed. If this occurs, it is a bug in Orville.

Since: 1.0.0.0

data ConnectionError

Orville throws a ConnectionError on an error reported by the underlying LibPQ connection that does not come directly from executing SQL. This could could represent an inability to open a new database connection, but could also represent other errors such as an error while quoting a database identifier.

Since: 1.0.0.0

data SqlExecutionError

Orville throws a SqlExecutionError when an error is reported by the underlying LibPQ connection during an attempt to execute SQL.

Since: 1.0.0.0

Constructors

SqlExecutionError

Fields

sqlExecutionErrorExecStatus :: Maybe ExecStatus

The underlying LibPQ execution status

sqlExecutionErrorMessage :: ByteString

Error message reported by PostgreSQL

sqlExecutionErrorSqlState :: Maybe ByteString

Any SQL state value reported by PostgreSQL. This can be used to programming determine what kind of error happening without needing to parse the error message. See https://www.postgresql.org/docs/current/errcodes-appendix.html

sqlExecutionErrorSqlQuery :: ByteString

The SQL query that was being run when the error occurred.

Re-exports from Data.Pool for convenience

data Pool a