Orville.PostgreSQL.Raw.PgTime

Description

Since: 1.0.0.0

Documentation

dayToPostgreSQL :: Day -> ByteString

Renders a Day value to a textual representation for PostgreSQL

Since: 1.0.0.0

day :: Parser Day

An Attoparsec parser for parsing Day from YYYY-MM-DD format. Parsing fails if given an invalid day.

Since: 1.0.0.0

utcTimeToPostgreSQL :: UTCTime -> ByteString

Renders a UTCTime value to a textual representation for PostgreSQL

Since: 1.0.0.0

utcTime :: Parser UTCTime

An Attoparsec parser for parsing UTCTime from an ISO 8601 style datetime and timezone with a few postgresql specific exceptions. See localTime for more details

Since: 1.0.0.0

localTimeToPostgreSQL :: LocalTime -> ByteString

Renders a 'Time.LocalTime value to a textual representation for PostgreSQL

Since: 1.0.0.0

localTime :: Parser LocalTime

An Attoparsec parser for parsing LocalTime from an ISO 8601 style datetime with a few exceptions. The seperator between the date and time is always ' ' and never T.

Since: 1.0.0.0