Orville.PostgreSQL.Raw.RawSql

Description

The funtions in this module are named with the intent that it is imported qualified as RawSql.

Since: 1.0.0.0

Documentation

data RawSql

RawSql provides a type for efficiently constructing raw sql statements from smaller parts and then executing them. It also supports using placeholder values to pass parameters with a query without having to interpolate them as part of the actual sql state and being exposed to sql injection.

Since: 1.0.0.0

parameter :: SqlValue -> RawSql

Includes an input parameter in the RawSql statement that will be passed using placeholders (e.g. '$1') rather than being included directly in the sql statement. This is the correct way to include input from untrusted sources as part of a RawSql query. The parameter must be formatted in a textual representation, which the database will interpret. The database type for the value will be inferred by the database based on its usage in the query.

Since: 1.0.0.0

fromString :: String -> RawSql

Constructs a RawSql from a String value using utf8 encoding.

Note that because the string is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

Since: 1.0.0.0

fromText :: Text -> RawSql

Constructs a RawSql from a Text value using utf8 encoding.

Note that because the text is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

Since: 1.0.0.0

fromBytes :: ByteString -> RawSql

Constructs a RawSql from a ByteString value, which is assumed to be encoded sensibly for the database to handle.

Note that because the string is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

Since: 1.0.0.0

intercalate :: (SqlExpression sql, Foldable f) => RawSql -> f sql -> RawSql

Concatenates a list of RawSql values using another RawSql value as the a separator between the items.

Since: 1.0.0.0

execute :: SqlExpression sql => Connection -> sql -> IO Result

Executes a RawSql value using the executeRaw function. Make sure to read the documentation of executeRaw for caveats and warnings. Use with caution.

Note that because this is done in IO no callback functions are available to be called.

Since: 1.0.0.0

executeVoid :: SqlExpression sql => Connection -> sql -> IO ()

Executes a RawSql value using the executeRawVoid function. Make sure to read the documentation of executeRawVoid for caveats and warnings. Use with caution.

Note that because this is done in IO no callback functions are available to be called.

Since: 1.0.0.0

connectionQuoting :: Connection -> Quoting IO

Quoting done in IO based using the quoting functions provided by the connection, which can apply quoting based on the specific connection properties.

If you don't have a connection available and are only planning on using the SQL for explanatory or example purposes, see exampleQuoting.

Since: 1.0.0.0

Fragments provided for convenience

space :: RawSql

Just a plain old space, provided for convenience

comma :: RawSql

Just a plain old comma, provided for convenience

commaSpace :: RawSql

Comma space separator, provided for convenience

leftParen :: RawSql

Just a plain old left paren, provided for convenience

rightParen :: RawSql

Just a plain old right paren, provided for convenience

dot :: RawSql

Just a plain period, provided for convenience

doubleQuote :: RawSql

Just a plain double quote, provided for convenience

doubleColon :: RawSql

Just two colons, provided for convenience

stringLiteral :: ByteString -> RawSql

Includes a bytestring value as string literal in the SQL statement. The string literal will be quoted and escaped for you, the value provided should not include surrounding quotes or quote special characters.

Note: It's better to use the parameter function where possible to pass values to be used as input to a SQL statement. There are some situations where PostgreSQL does not allow this, however (for instance, in some DDL statements). This function is provided for those situations.

Since: 1.0.0.0

identifier :: ByteString -> RawSql

Includes a bytestring value as an identifier in the SQL statement. The identifier will be quoted and escaped for you, the value provided should not include surrounding quotes or quote special characters.

Since: 1.0.0.0

parenthesized :: SqlExpression sql => sql -> RawSql

Constructs a RawSql by putting parentheses around an arbitrary expression. The result is returned as a RawSql. It is up to the caller to decide whether it should be wrapped in a more specific expression type.

Since: 1.0.0.0

Integer values as literals

intDecLiteral :: Int -> RawSql

Constructs a RawSql from an Int value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Since: 1.0.0.0

int8DecLiteral :: Int8 -> RawSql

Constructs a RawSql from an Int8 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Since: 1.0.0.0

int16DecLiteral :: Int16 -> RawSql

Constructs a RawSql from an Int16 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Since: 1.0.0.0

int32DecLiteral :: Int32 -> RawSql

Constructs a RawSql from an Int32 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Since: 1.0.0.0

int64DecLiteral :: Int64 -> RawSql

Constructs a RawSql from an Int64 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Since: 1.0.0.0

Generic interface for generating sql

class SqlExpression a where

SqlExpression provides a common interface for converting types to and from RawSql, either via toRawSql and unsafeFromRawSql, or the convenience function unsafeSqlExpression. Orville defines a large number of types that represent various fragments of SQL statements as well as functions to help construct the safely. These funtions can be found in the Expr. These types all provide SqlExpression instances as an escape hatch to allow you to pass any SQL you wish in place of what Orville directly supports. This should be use with great care as Orville cannot guarantee that the SQL you pass can be used to generate valid SQL in conjuction with the rest of the Expr API.

Since: 1.0.0.0

Methods

toRawSql :: a -> RawSql

unsafeFromRawSql :: RawSql -> a

unsafeSqlExpression :: SqlExpression a => String -> a

A conveinence function for creating an arbitrary SqlExpression from a String. Great care should be exercised in use of this function as it cannot provide any sort of guarantee that the string passed is usable to generate valid SQL via the rest of Orville's Expr API as the whatever SqlExpression type is returned.

For example, if one wanted build a boolean expression not support by Orville, you can do it like so

import qualified Orville.PostgreSQL.Expr as Expr

a :: Expr.BooleanExpr
a RawSql.unsafeSqlExpression "foo BETWEEN 1  AND 3"

Since: 1.0.0.0

toBytesAndParams :: (SqlExpression sql, Monad m) => Quoting m -> sql -> m (ByteString, [Maybe PgTextFormatValue])

Constructs the actual SQL bytestring and parameter values that will be passed to the database to execute a RawSql query. Any string literals thar are included in the SQL expression will be quoting using the given quoting directive.

Since: 1.0.0.0

toExampleBytes :: SqlExpression sql => sql -> ByteString

Builds the bytes that represent the raw sql. These bytes may not be executable on their own, because they may contain placeholders that must be filled in, but can be useful for inspecting sql queries.

Since: 1.0.0.0

data Quoting m

Provides procedures for quoting parts of a raw SQL query so that they can be safely executed. Quoting may be done in some Monad m, allowing for the use of quoting operations provided by Connection, which operates in the IO monad.

See connectionQuoting and exampleQuoting.

Since: 1.0.0.0

Constructors

Quoting

Fields

quoteStringLiteral :: ByteString -> m Builder

quoteIdentifier :: ByteString -> m Builder

exampleQuoting :: Quoting Identity

Quoting done in pure Haskell that is suitable for showing SQL examples, but is not guaranteed to be sufficient for all database connections. For quoting that is based on the actual connection to the database, see connectionQuoting.

Since: 1.0.0.0