Orville.PostgreSQL.Raw.SqlCommenter

Description

This module provides the very basics for sqlcommenter support.

Since: 1.0.0.0

Documentation

type SqlCommenterAttributes = Map Text Text

The representation of Text key/value pairs for supporting the sqlcommenter specification. This allows you to attach key/values of Text that supporting systems can use for advanced metrics. See sqlcommenter for details of the specification.

Since: 1.0.0.0

addSqlCommenterAttributes :: SqlExpression a => SqlCommenterAttributes -> a -> a

Adds a given SqlCommenter set of key/value Text pairs to a SqlExpression. This performs all of the required serialization for the given values. Note that no values are automatically added here, so any that you may wish to add can be freely set without a name clash of any kind from this function itself.

Since: 1.0.0.0