Orville.PostgreSQL.Raw.SqlValue

Description

The funtions in this module are named with the intent that it is imported qualified as SqlValue.

Since: 1.0.0.0

Documentation

data SqlValue

SqlValue represents a value that is in encoded format for use with LibPQ. It is used both for values passed to LibPQ and values parse from LibPQ. The conversions functions in Orville.PostgreSQL.Raw.SqlValue can be used to convert to and from the value.

Since: 1.0.0.0

isSqlNull :: SqlValue -> Bool

Checks whether the SqlValue represents a sql NULL value in the database.

Since: 1.0.0.0

sqlNull :: SqlValue

A value of SqlValue that will be interpreted as a sql NULL value when pasesed to the database.

Since: 1.0.0.0

fromInt8 :: Int8 -> SqlValue

Encodes an Int8 value for usage with database

Since: 1.0.0.0

toInt8 :: SqlValue -> Either String Int8

Attempts to decode a SqlValue as a Haskell Int8 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromInt16 :: Int16 -> SqlValue

Encodes an Int16 value for usage with database

Since: 1.0.0.0

toInt16 :: SqlValue -> Either String Int16

Attempts to decode a SqlValue as a Haskell Int16 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromInt32 :: Int32 -> SqlValue

Encodes an Int32 value for usage with database

Since: 1.0.0.0

toInt32 :: SqlValue -> Either String Int32

Attempts to decode a SqlValue as a Haskell Int32 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromInt64 :: Int64 -> SqlValue

Encodes an Int64 value for usage with database

Since: 1.0.0.0

toInt64 :: SqlValue -> Either String Int64

Attempts to decode a SqlValue as a Haskell Int value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromInt :: Int -> SqlValue

Encodes an Int value for usage with database

Since: 1.0.0.0

toInt :: SqlValue -> Either String Int

Attempts to decode a SqlValue as a Haskell Int value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromWord8 :: Word8 -> SqlValue

Encodes an Word8 value for usage with database

Since: 1.0.0.0

toWord8 :: SqlValue -> Either String Word8

Attempts to decode a SqlValue as a Haskell Word8 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromWord16 :: Word16 -> SqlValue

Encodes an Word16 value for usage with database

Since: 1.0.0.0

toWord16 :: SqlValue -> Either String Word16

Attempts to decode a SqlValue as a Haskell Word16 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromWord32 :: Word32 -> SqlValue

Encodes an Word32 value for usage with database

Since: 1.0.0.0

toWord32 :: SqlValue -> Either String Word32

Attempts to decode a SqlValue as a Haskell Word32 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromWord64 :: Word64 -> SqlValue

Encodes an Word64 value for usage with database

Since: 1.0.0.0

toWord64 :: SqlValue -> Either String Word64

Attempts to decode a SqlValue as a Haskell Word64 value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromWord :: Word -> SqlValue

Encodes an Word value for usage with database

Since: 1.0.0.0

toWord :: SqlValue -> Either String Word

Attempts to decode a SqlValue as a Haskell Word value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromDouble :: Double -> SqlValue

Encodes a Double value for usage with database

Since: 1.0.0.0

toDouble :: SqlValue -> Either String Double

Attempts to decode a SqlValue as a Haskell Double value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromBool :: Bool -> SqlValue

Encodes a Bool value for usage with database

Since: 1.0.0.0

toBool :: SqlValue -> Either String Bool

Attempts to decode a SqlValue as a Haskell Bool value. If decoding fails Nothing is returned.

Since: 1.0.0.0

fromText :: Text -> SqlValue

Encodes a Text value as utf8 so that it can be used with the database.

Since: 1.0.0.0

toText :: SqlValue -> Either String Text

Attempts to decode a SqlValue as UTF-8 text. If the decoding fails, Nothing is returned.

Note: This decoding _only_ fails if the bytes returned from the database are not a value UTF-8 sequence of bytes. Otherwise it always succeeds.

Since: 1.0.0.0

fromDay :: Day -> SqlValue

Encodes a Day value as text in YYYY-MM-DD format so that it can be used with the database.

Since: 1.0.0.0

toDay :: SqlValue -> Either String Day

Attempts to decode a SqlValue as into a Day value by parsing it from YYYY-MM-DD format. If the decoding fails Nothing is returned.

Since: 1.0.0.0

fromUTCTime :: UTCTime -> SqlValue

Encodes a UTCTime in ISO 8601 format for usage with the database.

Since: 1.0.0.0

toUTCTime :: SqlValue -> Either String UTCTime

Attempts to decode a SqlValue as a UTCTime formatted in iso8601 format with time zone. If the decoding fails, Nothing is returned.

Since: 1.0.0.0

fromLocalTime :: LocalTime -> SqlValue

Encodes a LocalTime in ISO 8601 format for usage with the database.

Since: 1.0.0.0

toLocalTime :: SqlValue -> Either String LocalTime

Attempts to decode a SqlValue as a LocalTime formatted in iso8601 format in the default Local. If the decoding fails, Nothing is returned.

Since: 1.0.0.0

fromRawBytes :: ByteString -> SqlValue

Creates a SqlValue from a raw byte string as if the bytes had returned by the database. This function does not interpret the bytes in any way, but the using decode functions on them might fail depending on whether the bytes can be parsed as the requested type.

Note: A value to represent a sql NULL cannot be constructed using this function. See fromRawBytesNullable for how to represent a nullable raw value.

Since: 1.0.0.0

fromRawBytesNullable :: Maybe ByteString -> SqlValue

Creates a SqlValue from a raw byte string. If Nothing is specified as the input parameter then the resulting SqlValue will represent a NULL value in sql. Otherwise the bytes given are used in the same way as fromRawBytes

Since: 1.0.0.0

toPgValue :: SqlValue -> Maybe PgTextFormatValue

Converts a SqlValue to its underlying raw bytes as it will be represented when sent to the database. The output should be recognizable as similar to to values you would write in query. If the value represents a sql NULL value, Nothing is returned

Since: 1.0.0.0