Orville.PostgreSQL.Schema.IndexDefinition

Description

Since: 1.0.0.0

Documentation

data IndexDefinition

Defines an index that can be added to a TableDefinition. Use one of the constructor functions below (such as uniqueIndex) to construct the index definition you wish to have and then use addTableIndexes. to add them to your table definition. Orville will then add the index next time you run auto-migrations.

Since: 1.0.0.0

indexCreationStrategy :: IndexDefinition -> IndexCreationStrategy

Gets the IndexCreationStrategy strategy to be used when creating the index described by the IndexDefinition. By default all indexes are created using the Transactional strategy.

Since: 1.0.0.0

setIndexCreationStrategy :: IndexCreationStrategy -> IndexDefinition -> IndexDefinition

Sets the IndexCreationStrategy strategy to be used when creating the index described by the IndexDefinition. By default all indexes are created using the Transactional strategy, but some tables are too large for for this to be feasible. See the Asynchronous creation strategy for how to work around this.

Since: 1.0.0.0

uniqueIndex :: NonEmpty FieldName -> IndexDefinition

Constructs an IndexDefinition for a UNIQUE index on the given columns.

Since: 1.0.0.0

uniqueNamedIndex :: String -> IndexBodyExpr -> IndexDefinition

Constructs an IndexDefinition for a UNIQUE index on the given columns.

Since: 1.0.0.0

nonUniqueIndex :: NonEmpty FieldName -> IndexDefinition

Constructs an IndexDefinition for a non-unique index on the given columns

Since: 1.0.0.0

nonUniqueNamedIndex :: String -> IndexBodyExpr -> IndexDefinition

Constructs an IndexDefinition for a non-unique index with given SQL and index name

Since: 1.0.0.0

mkIndexDefinition :: IndexUniqueness -> NonEmpty FieldName -> IndexDefinition

Constructs an IndexDefinition for an index on the given columns with the given uniquness.

Since: 1.0.0.0

mkNamedIndexDefinition :: IndexUniqueness -> String -> IndexBodyExpr -> IndexDefinition

Constructs an IndexDefinition for an index with the given uniqueness, given name, and given SQL.

Since: 1.0.0.0

data IndexUniqueness

Type to represent if an index should be unique.

Since: 1.0.0.0

Constructors

UniqueIndex

NonUniqueIndex

indexCreateExpr :: IndexDefinition -> Qualified TableName -> CreateIndexExpr

Gets the SQL expression that will be used to add the index to the specified table.

Since: 1.0.0.0

data IndexCreationStrategy

Defines how an IndexDefinition will be execute to add an index to a table. By default all indexes a created via the Transactional

Since: 1.0.0.0

Constructors

Transactional
The default strategy. The index will be added as part of the a database transaction along with all the other DDL being executed to migrate the database schema. If any migration should fail the index creation will be rolled back as part of the transaction. This is how schema migrations work in general in Orville.

Asynchronous
Creates the index asynchronously using the CONCURRENTLY keyword in PostgreSQL. Index creation will return immediately and the index will be created in the background by PostgreSQL. Index creation may fail when using the Asynchronous strategy. Orville has no special provision to detect or recover from this failure currently. You should manually check that index creation has succeeded. If necessary, you can manually drop the index to cause Orville to recreate it the next time migrations are run. This is useful when you need to add an index to a very large table because this can lock the table for a long time and cause an application to fail health checks at startup if migrations do not finish quickly enough. You should familiarize youself with how concurrent index creation works in PostgreSQL before using this. See https://www.postgresql.org/docs/current/sql-createindex.html#SQL-CREATEINDEX-CONCURRENTLY