Orville.PostgreSQL.Schema.SequenceIdentifier

Description

Since: 1.0.0.0

Documentation

data SequenceIdentifier

An identifier used by Orville to identify a particular sequence in a particular schema.

Since: 1.0.0.0

unqualifiedNameToSequenceId :: String -> SequenceIdentifier

Constructs a SequenceIdentifier where the sequence's name will not be qualified by a particular schema.

Since: 1.0.0.0

setSequenceIdSchema :: String -> SequenceIdentifier -> SequenceIdentifier

Sets the schema of the SequenceIdentifier. Wherever applicable, references to the sequence will be qualified by the given scheme name.

Since: 1.0.0.0

sequenceIdQualifiedName :: SequenceIdentifier -> Qualified SequenceName

Returns the 'Expr.Qualified Expr.SequenceName' that should be used to refer to the sequence in SQL queries.

Since: 1.0.0.0

sequenceIdUnqualifiedName :: SequenceIdentifier -> SequenceName

Returns the unqualified SequenceName that should be used to refer to the sequence in SQL queries where an unqualified reference is appropriate.

Since: 1.0.0.0

sequenceIdSchemaName :: SequenceIdentifier -> Maybe SchemaName

Returns the SchemaName (if any) that should be used to qualify references to the sequence in SQL queries.

Since: 1.0.0.0

sequenceIdToString :: SequenceIdentifier -> String

Converts a SequenceIdentifier for a string for descriptive purposes. The name will be qualified if a schema name has been set for the identifier.

Note: You should not use this function for building SQL expressions. Use sequenceIdQualifiedName instead for that.

Since: 1.0.0.0

sequenceIdUnqualifiedNameString :: SequenceIdentifier -> String

Retrieves the unqualified name of the sequence as a string.

Since: 1.0.0.0

sequenceIdSchemaNameString :: SequenceIdentifier -> Maybe String

Retrieves the schema name of the sequence as a string

Since: 1.0.0.0