Orville.PostgreSQL.Schema.TableIdentifier

Description

Since: 1.0.0.0

Documentation

data TableIdentifier

An identifier used by Orville to identify a particular table in a particular schema.

Since: 1.0.0.0

unqualifiedNameToTableId :: String -> TableIdentifier

Constructs a TableIdentifier where the table's name will not be qualified by a particular schema.

Since: 1.0.0.0

setTableIdSchema :: String -> TableIdentifier -> TableIdentifier

Sets the schema of the TableIdentifier. Wherever applicable, references to the table will be qualified by the given scheme name.

Since: 1.0.0.0

tableIdQualifiedName :: TableIdentifier -> Qualified TableName

Returns the 'Expr.Qualified Expr.TableName' that should be used to refer to the table in SQL queries.

Since: 1.0.0.0

tableIdUnqualifiedName :: TableIdentifier -> TableName

Returns the unqualified TableName that should be used to refer to the table in SQL queries where an unqualified reference is appropriate.

Since: 1.0.0.0

tableIdSchemaName :: TableIdentifier -> Maybe SchemaName

Returns the SchemaName (if any) that should be used to qualify references to the table in SQL queries.

Since: 1.0.0.0

tableIdToString :: TableIdentifier -> String

Converts a TableIdentifier for a string for descriptive purposes. The name will be qualified if a schema name has been set for the identifier.

Note: You should not use this function for building SQL expressions. Use tableIdQualifiedName instead for that.

Since: 1.0.0.0

tableIdUnqualifiedNameString :: TableIdentifier -> String

Retrieves the unqualified name of the table as a string.

Since: 1.0.0.0

tableIdSchemaNameString :: TableIdentifier -> Maybe String

Retrieves the schema name of the table as a string

Since: 1.0.0.0